Convert DataStore to typescript
