+++ src/bridge/IrcBridge.js
@@ -18,7 +18,7 @@ var IrcRoom = require("../models/IrcRoom");
 var IrcClientConfig = require("../models/IrcClientConfig");
 var BridgeRequest = require("../models/BridgeRequest");
 var stats = require("../config/stats");
-var DataStore = require("../DataStore");
+const { DataStore } = require("../DataStore");
 var log = require("../logging").get("IrcBridge");
 const {
     Bridge,
